pyproject.toml
[project]
name = "gemmasearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckduckgo-search>=8.0.1",
    "lxml>=5.4.0",
    "openai>=1.77.0",
    "playwright>=1.52.0",
    "pypdf2>=3.0.1",
    "readability-lxml>=0.8.4.1",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
